+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- Unnecessary `&self` reference in macro-generated code if self was already a reference.
 
 ## [0.4.11] - 2023-03-29
 ### Fixed